@@ -2,13 +2,13 @@
 /* Copyright Contributors to the ODPi Egeria project. */
 package org.odpi.openmetadata.http;
 
-import javax.servlet.Filter;
-import javax.servlet.FilterChain;
-import javax.servlet.FilterConfig;
-import javax.servlet.ServletException;
-import javax.servlet.ServletRequest;
-import javax.servlet.ServletResponse;
-import javax.servlet.http.HttpServletRequest;
+import jakarta.servlet.Filter;
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.FilterConfig;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.ServletRequest;
+import jakarta.servlet.ServletResponse;
+import jakarta.servlet.http.HttpServletRequest;
 import java.io.IOException;
 import java.util.HashMap;
 import java.util.List;
@@ -18,8 +18,9 @@ import java.util.Map;
  * HttpRequestHeadersFilter is a Java Servlet Filter that receives a list of http header names which are then
  * saved from the request in HttpHeadersThreadLocal
  */
+
 public class HttpRequestHeadersFilter implements Filter {
-    private List<String> headerNames;
+    private final List<String> headerNames;
 
     /**
      * @param headerNames list of http header names
