@@ -8,7 +8,8 @@ import org.odpi.openmetadata.userinterface.uichassis.springboot.auth.TokenUser;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.Authentication;
 
-import javax.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletRequest;
+import org.springframework.security.core.context.SecurityContextHolder;
 
 /**
  * This class provides a method to access the user name from the servlet session. This class should be subclassed so that the user
@@ -18,8 +19,8 @@ public class SecureController {
     public static final String PAGE_OFFSET_DEFAULT_VALUE = "0";
     public static final String PAGE_SIZE_DEFAULT_VALUE = "0";
 
-    @Autowired
-    private AuthService authService;
+//    @Autowired
+//    private AuthService authService;
 
     /**
      * Return user name if there is one or null. Passing null as the user to a rest call should result in a user not authorized error.
@@ -28,7 +29,7 @@ public class SecureController {
      */
     protected String getUser(HttpServletRequest request) {
         String userName = null;
-        Authentication auth = authService.getAuthentication(request);
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
         if(auth != null && auth.getDetails() != null && (auth.getDetails() instanceof TokenUser)){
             userName = ((TokenUser) auth.getDetails()).getUsername();
         }
